# docs/api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CAN-HMI Signal API",
    "description": "Real-time CAN bus signal monitoring and control API",
    "version": "0.1.0"
  },
  "paths": {
    "/signals": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "List latest signal values",
        "operationId": "list_signals_signals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/signals/available": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "List all available signals with metadata",
        "description": "Return the full metadata list for all signals.\n\nThe client calls this once at startup to get the structure, then only subscribes\nto lightweight value + timestamp updates over WebSocket.",
        "operationId": "list_available_signals_signals_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalMetadataListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/signals/{signal_name}": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "Get latest value for one signal",
        "operationId": "get_signal_signals__signal_name__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalValueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Signals"
        ],
        "summary": "Write value to signal (CAN write)",
        "operationId": "write_signal_signals__signal_name__put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteSignalRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/signals/batch_update": {
      "post": {
        "tags": [
          "Signals"
        ],
        "summary": "Write multiple writable signals simultaneously (batch)",
        "description": "Write multiple CAN signals at once.\n\nSignals belonging to the same CAN message are grouped together and sent as a single\nframe. Other signals in the same message that are not included in the\nbatch use ``writer.use_prevalue_for_unwritten_signal`` from ``system.json``.\nREST writes are broadcast immediately to all subscribed WS clients.",
        "operationId": "batch_update_signals_signals_batch_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSignalWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/config/signal/{signal_name}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Get config for one signal",
        "operationId": "get_signal_config_config_signal__signal_name__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devmode/signals": {
      "post": {
        "tags": [
          "Dev Mode"
        ],
        "summary": "Apply one signal family to several seats at once",
        "operationId": "apply_devmode_signal_api_devmode_signals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevModeSignalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Warning or access error code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Short description for the frontend"
          },
          "profile_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name",
            "description": "Profile name currently in effect"
          },
          "required_permission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Permission",
            "description": "Permission required for the operation"
          },
          "signal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Name",
            "description": "Affected signal name if this is a single-signal warning"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signals",
            "description": "List of signals skipped in the batch"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "AccessWarning",
        "description": "Access/profile warning information for a profile-scoped operation."
      },
...
      "BatchSignalWrite": {
        "properties": {
          "signals": {
            "items": {
              "$ref": "#/components/schemas/BatchSignalWriteItem"
            },
            "type": "array",
            "title": "Signals",
            "description": "List of signals to be written"
          }
        },
        "type": "object",
        "required": [
          "signals"
        ],
        "title": "BatchSignalWrite",
        "description": "Request to write multiple CAN signals at once — POST /signals/batch_update."
      },
...
      "DevModeSignalRequest": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Signal family: ACR_RetractRequest | ABL_RetractRequest | ISB_Color | HB_Request"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Value applied to every selected seat"
          },
          "seats": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Seats",
            "description": "Map seat_id → whether the value is applied"
          },
          "block_timeout_sec": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 3600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Timeout Sec",
            "description": "Seat lock renewal duration (seconds); defaults to 60"
          }
        },
        "type": "object",
        "required": [
          "signal_name",
          "value",
          "seats"
        ],
        "title": "DevModeSignalRequest",
        "description": "Apply one signal family to several seats at once — POST /api/devmode/signals."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignalConfigResponse": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Signal identifier"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Measurement unit"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value",
            "description": "Minimum valid value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value",
            "description": "Maximum valid value"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name",
            "description": "Functional grouping"
          },
          "widget_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Type",
            "description": "Frontend widget type"
          },
          "writable": {
            "type": "boolean",
            "title": "Writable",
            "description": "Whether the signal is writable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "signal_name"
        ],
        "title": "SignalConfigResponse",
        "description": "Display configuration and metadata for a signal."
      },
      "SignalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SignalValueResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of signals"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of signals in the list"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/AccessWarning"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Access warnings, if any"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SignalListResponse",
        "description": "List of signal values returned by the API."
      },
      "SignalMetadata": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Unique signal identifier"
          },
          "std_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Std Name",
            "description": "Standard signal name; currently identical to signal_name"
          },
          "tag": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "Tags inferred from the signal name or DBC configuration"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Measurement unit"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value",
            "description": "Minimum valid value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value",
            "description": "Maximum valid value"
          },
          "writable": {
            "type": "boolean",
            "title": "Writable",
            "description": "Whether the signal can be written via the API",
            "default": false
          },
          "states": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "States",
            "description": "List of enum states [{value, description}], or None for continuous numeric signals"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name",
            "description": "Functional group (for example: engine, body)"
          },
          "widget_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Type",
            "description": "Frontend widget type"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Current value (snapshot, optional)"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Unix timestamp of the latest read"
          }
        },
        "type": "object",
        "required": [
          "signal_name"
        ],
        "title": "SignalMetadata",
        "description": "Full metadata for one signal — returned by GET /signals/available."
      },
      "SignalMetadataListResponse": {
        "properties": {
          "signals_info": {
            "items": {
              "$ref": "#/components/schemas/SignalMetadata"
            },
            "type": "array",
            "title": "Signals Info",
            "description": "List of signal metadata"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of signals"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/AccessWarning"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Access warnings, if any"
          }
        },
        "type": "object",
        "required": [
          "signals_info",
          "total"
        ],
        "title": "SignalMetadataListResponse",
        "description": "List of signal metadata."
      },
      "SignalValueResponse": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Unique signal identifier"
          },
          "std_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Std Name",
            "description": "Standard signal name; currently identical to signal_name"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Decoded real value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Measurement unit (for example: km/h, °C)"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp",
            "description": "Unix timestamp (seconds) when the value was read"
          }
        },
        "type": "object",
        "required": [
          "signal_name",
          "value",
          "timestamp"
        ],
        "title": "SignalValueResponse",
        "description": "Current value of a CAN signal."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WriteSignalRequest": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Value to write to the CAN bus"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "WriteSignalRequest",
        "description": "Request to write a value onto the CAN bus."
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}